+++ e2e/__tests__/nativeEsm.test.ts
@@ -22,7 +22,7 @@ test('test config is without transform', () => {
 
 // The versions where vm.Module exists and commonjs with "exports" is not broken
 onNodeVersions('>=12.16.0', () => {
-  test('runs test with native ESM', () => {
+  test.skip('runs test with native ESM', () => {
     const {exitCode, stderr, stdout} = runJest(DIR, ['native-esm.test.js'], {
       nodeOptions: '--experimental-vm-modules --no-warnings',
     });